backend/pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Backend for the LangGraph agent"
authors = [
    { name = "Philipp Schmid", email = "schmidphilipp1995@gmail.com" },
]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.9,<4.0"
dependencies = [
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0", 
    "langchain-anthropic>=0.1.20",
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langgraph>=0.1.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "aiohttp>=3.8.0",
    "beautifulsoup4>=4.12.0",
    "instructor>=1.4.3",
    "lxml>=4.9.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "google-generativeai>=0.3.0",
    "anthropic>=0.25.0",
    "openai>=1.0.0",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
